+++ reversiAPI/players/models/models.py
@@ -5,7 +5,7 @@ from torch import nn
 class ModelDqn1(nn.Module):
 
     def __init__(self, n_in, n_mid, n_out):
-        super(TheModelClass1, self).__init__()
+        super(ModelDqn1, self).__init__()
         self.fc1 = nn.Linear(n_in, n_mid)
         self.fc2 = nn.Linear(n_mid, n_mid)
         self.fc3 = nn.Linear(n_mid, n_out)
@@ -20,7 +20,7 @@ class ModelDqn1(nn.Module):
 class ModelDqn2(nn.Module):
 
     def __init__(self, n_in, n_mid, n_out):
-        super(TheModelClass2, self).__init__()
+        super(ModelDqn2, self).__init__()
         self.fc1 = nn.Linear(n_in, n_mid)
         self.fc2 = nn.Linear(n_mid, n_mid)
         self.fc3 = nn.Linear(n_mid, n_mid)
@@ -43,12 +43,12 @@ class ModelDqn2(nn.Module):
 class ModelSl1(nn.Module):
 
     def __init__(self):
-        super(TheModelClassSl1, self).__init__()
+        super(ModelSl1, self).__init__()
 
         self.conv1 = nn.Conv2d(1, 32, 3, padding=1)
         self.conv2 = nn.Conv2d(32, 64, 3, padding=1)
 
-        self.fc1 = nn.Linear(64*4*4, 128)
+        self.fc1 = nn.Linear(64 * 4 * 4, 128)
         self.fc2 = nn.Linear(128, 64)
         self.fc3 = nn.Linear(64, 64)
 
@@ -57,7 +57,7 @@ class ModelSl1(nn.Module):
         x = F.relu(self.conv2(x))
         x = F.max_pool2d(x, (2, 2))
 
-        x = x.view(-1, 64*4*4)
+        x = x.view(-1, 64 * 4 * 4)
 
         x = F.relu(self.fc1(x))
         x = F.relu(self.fc2(x))
@@ -65,25 +65,25 @@ class ModelSl1(nn.Module):
 
         return output
 
+
 class ModelSl2(nn.Module):
 
     def __init__(self):
-        super(TheModelClassSl2, self).__init__()
+        super(ModelSl2, self).__init__()
 
         self.conv1 = nn.Conv2d(2, 32, 3, padding=1)
         self.conv2 = nn.Conv2d(32, 64, 3, padding=1)
 
-        self.fc1 = nn.Linear(64*4*4, 128)
+        self.fc1 = nn.Linear(64 * 4 * 4, 128)
         self.fc2 = nn.Linear(128, 64)
         self.fc3 = nn.Linear(64, 64)
-
 
     def forward(self, x):
         x = F.relu(self.conv1(x))
         x = F.relu(self.conv2(x))
         x = F.max_pool2d(x, (2, 2))
 
-        x = x.view(-1, 64*4*4)
+        x = x.view(-1, 64 * 4 * 4)
 
         x = F.relu(self.fc1(x))
         x = F.relu(self.fc2(x))
